Moved ResourceType enum into metadata package and specifying how content models map to resource types in the enum
